refactor: some prefix tests
