+++ UNSAFE.md
@@ -24,6 +24,11 @@ cd fuzz/compare_prefixed_slice
 mkdir in
 cat /dev/random | head -n 100 > in/input
 cargo afl build && cargo afl fuzz -i in -o out target/debug/compare_prefixed_slice
+
+cd fuzz/prefix_filter
+mkdir in
+cat /dev/random | head -n 100 > in/input
+cargo afl build && cargo afl fuzz -i in -o out target/debug/prefix_filter
 ```
 
 ## Run mutation testing
